Add Exple 002

+++ homeWork06/Program.cs
@@ -1,6 +1,6 @@
 ﻿Console.WriteLine("Введите число ");
 
-double a = Convert.ToInt32(Console.ReadLine());
+int a = Convert.ToInt32(Console.ReadLine());
 
 Console.WriteLine(a % 2 == 0 ? "Четное" : "Не четное");
